break up big test framework, clean up naming infra
